ReviewReport01
Review report on Exercises 01 report by Spencer Woody 

The report includes solutions to theoretical questions, simulation results and thoughts with R code attached. For linear regression section, detailed steps to derive $\hat \beta$ are provided for part (A). Then, Cholesky decomposition method is proposed to calculate WLS solutions. The inversion method and Cholesky method are benchmarked with simulated data. For the last part, the R package Matrix is used to exploit the sparsity property. Benchmarking is implemented across different parameter settings. For generalized linear models section, the gradient descent method and the Newton’s method are implemented with results properly demonstrated. Thoughts on comparison of two methods are given in the last part.

The report is complete and well written with a nice structure and flow. The mathematical derivations are correct and written clearly. Most results are shown in a proper way with explanation. The simulation is implemented in a careful and cautious way. 

Overall, I like this report. Some suggestions for improvement are as follows:

1.	During the class, a student shared her thoughts about different decomposition methods including LU decomposition and QR decomposition. If you took notes somewhere else, please ignore this suggestion. Otherwise, I think it will be helpful to aggregate those into the report. Implementing them in R and benchmarking to see the tradeoffs in using different methods.
2.	To show the results from benchmarking two methods, screenshots of the console in R are provided. This is definitely an easy and direct way to demonstrate results to those who are familiar with R. However, I think a better way is to list results in a table with some explanation or plot the results in figures. 
3.	For the code part, I only have a minor suggestion. In line 17, the matrix W is defined as a diagonal matrix. Thus, it is not necessary to diagonalize it in line 22 and 23 when defining function Inv.method.